Stability and better error reporting
This was an investigation and attempt to fix issue #30.

It turns out there wasn't actually much to fix, the real issue ended up being a bug in electron version v0.33.1-2 (and possibly some lower versions) desribed in patch notes:

Fix length of Buffer becoming 0 when there are lots of Buffer created.
Fix crash when there is external ArrayBuffer created with pointers to addresses outside the heap.
After updating to latest version of electron the updater started working. This patch actually includes a lot of error propagation and logging improvements however. So I want to wrap this up and get it out.

Sending as a PR for now, in case anyone wants to review or comment. I will merge and publish after a few other fixes.