Emit an error event in addition to logging
This fix adds an error event to the updater which can be called by the start function.

Additionally errors are now logged and emitted properly. Also callback is only called when appropriate.

A user may want to handle certain kinds of errors manually, such as
providing a UI for entering proxy information before restarting or
showing a custom error reporting UI.